requirements.txt
# core utils
numpy
scipy
pandas
scikit-learn
tqdm
matplotlib
seaborn
pyyaml

# EEG + ML
mne
pyEDFlib
einops
omegaconf
hydra-core
pytorch-lightning
optuna
wandb
tensorboard
python-dotenv